Added polling on set interval and constructor on TMEvent

+++ src/schema.rs
@@ -4,13 +4,13 @@ use chrono::{DateTime, Local};
 #[derive(Debug)]
 pub struct TMEvent {
     pub id: String,
-    pub num_offers: u32,
+    pub num_offers: usize,
     pub last_updated: DateTime<Local>,
 }
 
 impl TMEvent {
 
-    pub fn new(a_id: String, a_num_offers: u32) -> Self {
+    pub fn new(a_id: String, a_num_offers: usize) -> Self {
 
         let new_self = Self {
             id: a_id,
